fix: update

+++ app/Services/UserService.php
@@ -2,27 +2,28 @@
 
 namespace App\Services;
 
-use App\Http\Requests\NewsFormRequest;
-use App\Models\News;
+use App\DTO\UserDTO;
 use App\Models\User;
-use Carbon\Carbon;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Facades\Crypt;
-use Illuminate\Support\Facades\Log;
 
 final class UserService
 {
-    public function getAll(int $perPage):LengthAwarePaginator
+    public function getAll(int $perPage): LengthAwarePaginator
     {
         return User::paginate($perPage);
     }
 
     public function getBySearching(string $search)
     {
-        return User::where('name', 'like', "%$search%")->orWhere('email', 'like', "%$search%")->paginate(12)->withQueryString();
+        return User::where(function ($query) use ($search): void {
+            $query->where('name', 'like', "%$search%")
+                ->orWhere('email', 'like', "%$search%");
+        })->paginate(12)->withQueryString();
     }
 
-    public function getById(int $id){
+    public function getById(int $id)
+    {
         return User::where('id', '=', $id)->select(['name', 'email', 'isAdmin', 'id'])->first();
     }
 
@@ -31,24 +32,31 @@ final class UserService
         return Crypt::decrypt(User::where('id', '=', $id)->pluck('formula')->first());
     }
 
-    public function destroy(int $id):bool
+    public function destroy(int $id): bool
     {
         $user = User::find($id);
+        if (! $user) {
+            return false;
+        }
+
         return $user->delete();
     }
 
-    public function approveUser(int $id):bool{
+    public function approveUser(int $id): bool
+    {
         $user = User::find($id);
-        if ($user){
+        if ($user) {
             $user->approved = 1;
             $user->save();
+
             return true;
         }
+
         return false;
     }
 
-    public function update(int $userId, string $formula):bool
+    public function update(UserDTO $dto): bool
     {
-        return User::where('id', '=', $userId)->update(['formula' => Crypt::encrypt($formula)]);
+        return User::where('id', '=', $dto->userId)->update(['formula' => Crypt::encrypt($dto->formula)]);
     }
 }
